+++ packages/ui/Routable/Link.js
@@ -1,0 +1,69 @@
+import kind from '@enact/core/kind';
+import {forProp, forward, handle, preventDefault} from '@enact/core/handle';
+import hoc from '@enact/core/hoc';
+import PropTypes from 'prop-types';
+import React from 'react';
+
+import {RouteContext} from './util';
+
+const LinkBase = kind({
+	name: 'Link',
+
+	propTypes: {
+		path: PropTypes.string.isRequired,
+		disabled: PropTypes.bool
+	},
+
+	defaultProps: {
+		disabled: false
+	},
+
+	handlers: {
+		onClick: handle(
+			forProp('disabled', false),
+			forward('onClick'),
+			preventDefault
+		)
+	},
+
+	render: ({path, ...rest}) => {
+		return (
+			<a href={path} {...rest} />
+		);
+	}
+});
+
+const Linkable = hoc({navigate: 'onClick'}, (config, Wrapped) => {
+	const {navigate} = config;
+
+	return kind({
+		name: 'Linkable',
+
+		contextType: RouteContext,
+
+		propTypes: {
+			path: PropTypes.string.isRequired
+		},
+
+		handlers: {
+			[navigate]: (ev, {path}, {navigate: nav}) => {
+				if (nav) nav({path});
+			}
+		},
+
+		render: (props) => {
+			return (
+				<Wrapped {...props} />
+			);
+		}
+	});
+});
+
+const Link = Linkable(LinkBase);
+
+export default Link;
+export {
+	Link,
+	LinkBase,
+	Linkable
+};